Change a lot of thing (#10)
* Change a lot of thing
+ Change the way create Puzzle (now it is sure that have the way)
+ Fix gameLogic so that food cannot eat wall (not sure enough)
+ KeyPress now return value so that can draw background when need
+ Point now is float
+ Change something in constant

+++ Food.cpp
@@ -5,7 +5,7 @@ void Food::Random()
 	Obstacle::Random();
 	_text = '0';
 }
-void Food::CollidePoccess(int& point, Ball& ball)
+void Food::CollidePoccess(float& point, Ball& ball)
 {
 	Obstacle::CollidePoccess(point, ball);
