+++ Food.h
@@ -6,7 +6,7 @@ class Food : public Obstacle
 {
 public:
 	void Random();
-	void CollidePoccess(int& point, Ball& ball);
+	void CollidePoccess(float& point, Ball& ball);
 
 public:
 	Food();
